Extracts CreateNewEventDialog
