+++ lib/mutations/event-types/create-event-type.ts
@@ -1,10 +1,12 @@
-import { EventType } from "@prisma/client";
-
 import * as fetch from "@lib/core/http/fetch-wrapper";
-import { CreateEventType } from "@lib/types/event-type";
+import { CreateEventType, CreateEventTypeResponse } from "@lib/types/event-type";
 
 const createEventType = async (data: CreateEventType) => {
-  const response = await fetch.post<CreateEventType, EventType>("/api/availability/eventtype", data);
+  const response = await fetch.post<CreateEventType, CreateEventTypeResponse>(
+    "/api/availability/eventtype",
+    data
+  );
+  console.log(`response`, response);
   return response;
 };
